Split mainApp into App and GaugePage
To improve the ability to add other pages to the app later I have taken the widgets which make up the contents of the 'Gauges page' to their own separate file.

The app page will be used as the main, window, and will show the contents it is provided. This was I can add another page in future like a controls page